extract func usage() for easier navigation
